requirements.txt
bcrypt==3.2.0
Flask==2.0.1
Flask-Session==0.4.0
pymongo==3.12.0
dnspython==1.16.0
Werkzeug==2.0.1
gunicorn==20.1.0
pyfiglet==0.8.post1
requests==2.26.0
opencv-python==4.5.3.56
Pillow==8.3.1


